Rename js file to controll database (creating Clients)
