+++ routes/friendsList.js
@@ -1,4 +1,4 @@
-const { createClientDB } = require("./test_db");
+const { createClientDB } = require("./databaseController");
 
 // =====================================================================================================
 // PUBLIC:
